+++ src/librustc_trans/trans/context.rs
@@ -510,7 +510,7 @@ impl<'tcx> LocalCrateContext<'tcx> {
         CrateContext {
             shared: shared,
             local: self,
-            index: !0 as usize,
+            index: !0,
         }
     }
 }
